Implemented evasion/counter system. Fixed some bugs. Also changed the directory structure of the autogenerated directories for the mod.
Evasion system can be modified via the xml autogenerated in the TurnBasedMinecraft folder.
Details on what the variables do can be seen in the code, in the xml, and in this commit message.

evasionRate: default evasion rate for a combatant.
criticalRate: default critical hit rate for a combatant.
onDodgeEvasionRate: evasion rate when dodging an attack from a combatant that wasn't selected.
onCorrectDodgeEvasionRate: evasion rate when dodging an attack from a combatant that was selected.
counterRateAfterHit: counter-attack rate after being hit by a combatant.
counterRateAfterMiss: counter-attack rate after dodging an attack by a combatant.
onCorrectDodgeHitBonus: hit bonus that increases the chances of hitting a combatant on the next turn.
onCorrectDodgeCriticalBonus: critical bonus that increases the chances of a critical hit on a combatant on the next turn.